Move get layer names func to utils
